增加重连逻辑，只要对端关闭直接关闭当前socket并重新监听

+++ sim_lib.h
@@ -33,7 +33,7 @@ struct ResultStruct
 {
 	int idx;
 	int arg_int[3];
-	float arg_float[4];
+	double arg_float[4];
 };
 
 int main_run(float* config, int idx, ResultStruct* result);
